pyproject.toml
[tool.poetry]
name = "CostsMap"
version = "2.0.0"
description = "CostsMap is a site for costs accounting"
authors = ["artemowkin <artyom.groshev2017@gmail.com>"]
license = "GPL-3.0"

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.4"
djangorestframework = "^3.12.4"
dj-database-url = "^0.5.0"
django-allauth = "^0.44.0"
gunicorn = "^20.1.0"
psycopg2-binary = "^2.8.6"
whitenoise = "^5.2.0"
python-dateutil = "^2.8.1"
django-service-objects = "^0.7.0"
simplejson = "^3.17.2"
dj-rest-auth = "^2.1.7"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
